requirements.txt
---
torch==2.0.1
torchaudio==2.0.2
transformers==4.35.0
librosa==0.10.0
openai-whisper==20231106
pysrt==1.1.2
numpy==1.24.3
pandas==2.0.3
tqdm==4.66.1
matplotlib==3.7.1
sacrebleu==2.3.1
jiwer==2.8.0
---
